requirements.txt
fastapi==0.135.1
sqlmodel==0.0.37
SQLAlchemy==2.0.48
uvicorn[standard]==0.41.0
python-jose==3.5.0
pyasn1==0.6.2
psycopg[binary]==3.2.12
